Migrate to TrustInterface exposing continuous trust levels on [0, 1.0].

+++ agent/src/agent/trust/RandomTrustNetwork.java
@@ -234,7 +234,7 @@ public class RandomTrustNetwork extends TrustNetwork {
 		Context c = new Context(era.getName());
 		Peer p = new Peer(agent);
         
-        trust.storeRecommendation(c, p, (int)certainty);
+        trust.storeRecommendation(c, p, certainty);
 	}
 
 	/* (non-Javadoc)
